Add files via upload
https://www.theepochtimes.com/epochfun/sudoku-hard-central-4087017
Sudoku is a logic-based puzzle game that is played on a 9x9 grid. The main goal of the game is to fill the grid with numbers from 1 to 9, ensuring that each row, column, and 3x3 subsection of the grid contains unique numbers.

Sudoku is a mental and logical game that requires concentration and logical thinking to solve. It is beneficial for strengthening analytical and numerical skills, as well as improving memory and focus.

By solving Sudoku, you engage your brain in problem-solving exercises, which can help in enhancing your abilities in planning, problem-solving, and decision-making.